Bug 1590619: Only clean up Remote Canvas IPC classes when the CanvasChild owns the only recorder reference. r=jrmuizel
This also removes the DrawEventRecorderPrivate::IsEmpty method as it is no
longer used. It doesn't undo the splitting out of the tracking of the
UnscaledFonts, which was added at the same time, because I think it is useful
to highlight the fact that they are never currently removed.

Differential Revision: https://phabricator.services.mozilla.com/D51593

--HG--
extra : moz-landing-system : lando