Bug 1655476 - Simplify including AOSP headers in b2g builds r=glandium,gerard-majax
This patch removes all the manual includes of AOSP headers in the build
system by moving all the headers into the sysroot and only including
that instead. Some of the #include directives required slight
adjustments but this makes the whole source more consistent. The
relevant TaskCluster configuration is adjusted to pick up this changes
and the manifest is updated to the latest version.

Differential Revision: https://phabricator.services.mozilla.com/D87055