Bug 1676417 - fix use of 2020h2 bookmarks pref in the default location pref getter in PlacesUIUtils to unbreak beta tests, r=jaws
The browser_autoshow_bookmarks_toolbar.js test pushPrefEnv's the 2020h2 pref,
but because the getter for the default location of bookmarks caches the pref
value for the default location, and doesn't update when the 2020h2 pref
changes, we continue returning the 'unfiled' folder from the pref getter if
the getter has been previously accessed (e.g. when the tests are run in a
directory).

To fix this, decouple the check for the 2020h2 pref from the preference
getter for the default location.

Differential Revision: https://phabricator.services.mozilla.com/D96592