Bug 1674224: Generate Java signature hints in JNI wrappers; r=geckoview-reviewers,agi
This patch adds some code to our annotation processor that produces comments in
our generated JNI headers to help the developer understand which types are
involved. There are two new types of comments:

* For generated methods from native to Java, we produce a simplified Java method
  signature (sans package names) to help the caller disambiguate parameter
  types. This is helpful since the machine-readable method signature tend to
  use `mozilla::jni::Object::Param` instead of actual types (for the purposes of
  avoiding dependency hell, presumably).
* For Java to native methods, we produce a sample signature that is usable for
  the implementation of the native method.

While we did discuss some ideas about other locations where we could output this
data, I'd like to start with the generated header files for now. We can make
adjustments to the output location in future bugs.

I will be attaching an example of a generated header that contains these new
comments to the bug.

Differential Revision: https://phabricator.services.mozilla.com/D96236