Bug 111994 - Properly re-attach delegates after a daemon restart. r=seinlin.li
Only the power manager delegate was implementing maybe_xpcom() which is not
needed by all session objects but is by the ones backed by an xpcom component.

This patch refactors session object management by splitting the Xpcom support
out to make it clear what to implement on each kind of session object.